* libc/minires.c (scanline): Accommodate ctype changes which disallow use of an unadorned char argument to is* macros. * regex/regcomp.c: Ditto, throughout. * regex/regex2.h (ISWORD): Ditto.

+++ winsup/cygwin/regex/regex2.h
@@ -131,4 +131,4 @@ struct re_guts {
 
 /* misc utilities */
 #define	OUT	(CHAR_MAX+1)	/* a non-character value */
-#define	ISWORD(c)	(isalnum(c) || (c) == '_')
+#define	ISWORD(c)	(isalnum((unsigned)c) || (c) == '_')
